update_10

+++ lib/Fierbase/controllers/user_fb_controller.dart
@@ -1,4 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import '../../enums.dart';
 import '../../Models/fb/user_model.dart';
 import '../../Fierbase/controllers/fb_storage_controller.dart';
 
@@ -49,18 +50,31 @@ class UserFbController {
         .snapshots();
   }
 
-  Future<UserModel?> showUser(UserModel user) async {
-    var result = await _firestore
-        .collection(_collection)
-        .doc(user.id)
-        .withConverter<UserModel>(
-          fromFirestore: (snapshot, options) =>
-              UserModel.fromJson(snapshot.data()!),
-          toFirestore: (value, options) => value.toJson(),
-        )
-        .get();
 
-    return result.data();
+  Stream<QuerySnapshot<UserModel>> showVendor() async* {
+    yield* _firestore
+        .collection(_collection)
+        .where('userType', isEqualTo: UserType.vendor.name)
+        .orderBy('timestamp', descending: true)
+        .withConverter<UserModel>(
+      fromFirestore: (snapshot, options) =>
+          UserModel.fromJson(snapshot.data()!),
+      toFirestore: (value, options) => value.toJson(),
+    )
+        .snapshots();
+  }
+
+  Stream<QuerySnapshot<UserModel>> search(String keyword) async* {
+    yield* _firestore
+        .collection(_collection)
+        .where('name', isGreaterThanOrEqualTo: keyword)
+        .withConverter<UserModel>(
+      fromFirestore: (snapshot, options) =>
+          UserModel.fromJson(snapshot.data()!),
+      toFirestore: (value, options) => value.toJson(),
+    )
+        .orderBy('timestamp', descending: true)
+        .snapshots();
   }
 
   Future<bool> updateUser(UserModel user) async {
@@ -75,8 +89,14 @@ class UserFbController {
   }
 
   Future<void> deleteUser(UserModel user) async {
+    // Delete user's data from Firestore
     await _firestore.collection(_collection).doc(user.id).delete();
-    await FbStorageController().deleteFile(user.profileImg!.path!);
+    if(user.profileImg != null){
+      await FbStorageController().deleteFile(user.profileImg!.path!);
+    }
+    // Delete the user's authentication account
+    //await FbAuthController().deleteAccount(user.id ?? '');
+
   }
 }
